Change build directory structure
Directory named 'build' was a directory for transpiled server codes.
In previous, utils are inside of server directory and I didn't used
modules of utils in server side, so the problem didn't occured.
But as I separate two directories and use isEmpty of nullChecker.js, I had
to change the build directory structure.

I made build/server and build/utils.
And because of this, I had to modify the relative path of 'public' path in
server/main.js.